+++ src/demux/mod.rs
@@ -12,7 +12,7 @@ mod wav;
 
 pub use mkv::MkvDemuxer;
 pub use mp4::Mp4Demuxer;
-pub use mux::{MkvMuxer, Mp4Muxer, MuxConfig, Muxer, OggMuxer, WavMuxer};
+pub use mux::{MkvMuxer, Mp4Muxer, MuxConfig, Muxer, OggMuxer, VideoMuxConfig, WavMuxer};
 pub use ogg::OggDemuxer;
 pub use wav::WavDemuxer;